Added all Quarry controls
Tutorialising these panels prevents the mechanics attached to them from working. In effect, the mill lift and ramp cannot be raised, the quarry elevator cannot be lowered, the boathouse ramp can't be raised or moved, and the boathouse hook can't be moved. Despite this, all panels in both areas are still possible to solve using snipes. Five EP's are disabled, however.