Added most Swamp controls
The sliding bridge can only be moved to the right (although it can be returned to the left simply by standing there, which makes the above ground leftward EP possible). This does make the underwater sliding bridge EP impossible.

The floodgate can only be closed, not opened, which makes two of the sand EP's impossible, and while the inward one starts out possible, if the floodgate is closed it can never be reopened which means the EP becomes impossible.

The rotating bridge was left alone because the default behaviour is to move CW to purple/red, which could result in the player getting softlocked.